The changes are aimed at readability of the project
The debug == False check is introduced everywhere.
This means,  to see the output from a function, extra parameter
debug = True should be supplied to call it

The tests were split into two groups: group that passes
and the group that does not  pass, for various reasons.
For usability of the projects, problems with the ones that do
not pass have to be figured out and resolved.

Possible causes of those problems involve but are not limited to:
* Fixes of missing edges ( most likely)

Signed-off-by: Georgiy Krylov <georgiykrylov@gmail.com>